Invoke InstantPage::Delegate functions only if the IPC message is for the active page entry and if the InstantPage should process the message.
BUG=none
TEST=none

Review URL: https://chromiumcodereview.appspot.com/16345002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@205117 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 72238aac3b0551c82d5908723042025f046986fe

+++ chrome/browser/ui/search/instant_page.cc
@@ -165,6 +165,18 @@ bool InstantPage::ShouldProcessNavigateToURL() {
   return false;
 }
 
+bool InstantPage::ShouldProcessDeleteMostVisitedItem() {
+  return false;
+}
+
+bool InstantPage::ShouldProcessUndoMostVisitedDeletion() {
+  return false;
+}
+
+bool InstantPage::ShouldProcessUndoAllMostVisitedDeletions() {
+  return false;
+}
+
 void InstantPage::RenderViewCreated(content::RenderViewHost* render_view_host) {
   if (ShouldProcessRenderViewCreated())
     delegate_->InstantPageRenderViewCreated(contents());
@@ -307,14 +319,35 @@ void InstantPage::OnSearchBoxNavigate(int page_id,
       contents(), url, transition, disposition, is_search_type);
 }
 
-void InstantPage::OnDeleteMostVisitedItem(const GURL& url) {
+void InstantPage::OnDeleteMostVisitedItem(int page_id, const GURL& url) {
+  if (!contents()->IsActiveEntry(page_id))
+    return;
+
+  OnInstantSupportDetermined(page_id, true);
+  if (!ShouldProcessDeleteMostVisitedItem())
+    return;
+
   delegate_->DeleteMostVisitedItem(url);
 }
 
-void InstantPage::OnUndoMostVisitedDeletion(const GURL& url) {
+void InstantPage::OnUndoMostVisitedDeletion(int page_id, const GURL& url) {
+  if (!contents()->IsActiveEntry(page_id))
+    return;
+
+  OnInstantSupportDetermined(page_id, true);
+  if (!ShouldProcessUndoMostVisitedDeletion())
+    return;
+
   delegate_->UndoMostVisitedDeletion(url);
 }
 
-void InstantPage::OnUndoAllMostVisitedDeletions() {
+void InstantPage::OnUndoAllMostVisitedDeletions(int page_id) {
+  if (!contents()->IsActiveEntry(page_id))
+    return;
+
+  OnInstantSupportDetermined(page_id, true);
+  if (!ShouldProcessUndoAllMostVisitedDeletions())
+    return;
+
   delegate_->UndoAllMostVisitedDeletions();
 }
